+++ conjur_api/models/hostfactory/create_host_data.py
@@ -19,9 +19,11 @@ class CreateHostData:
     def __init__(self,
                  # using id shadows the internal id
                  host_id: str = "",
-                 token: str = ""):
+                 token: str = "",
+                 annotations: dict = None):
         self.host_id = host_id
         self.token = token
+        self.annotations = annotations or {}
 
         if self.host_id.strip() == "":
             raise MissingRequiredParameterException("Missing required parameter, 'host_id'")
@@ -29,7 +31,7 @@ class CreateHostData:
         if self.token == "":
             raise MissingRequiredParameterException("Missing required parameter, 'token'")
 
-    def get_host_id(self):
+    def get_host_id(self) -> dict:
         """
         to_dict
 
@@ -41,5 +43,13 @@ class CreateHostData:
 
         return params
 
+    def get_annotations(self) -> dict:
+        """
+        :return: dictionary containing annotations in a format acceptable by Conjur REST API
+        """
+        return {
+            f"annotations[{key}]": value for key, value in self.annotations.items()
+        }
+
     def __repr__(self) -> str:
-        return f"{{'id': '{self.host_id}'"
+        return f"{{'id': '{self.host_id}', 'annotations': '{self.annotations}'"